refactor: address review — move impl out of __init__, fix imports and naming
- Move implementation from __init__.py to preshuffle_gemm.py (follows
  triton convention); __init__.py now just re-exports
- Merge _configs.py into preshuffle_gemm.py (config is kernel-specific)
- Use top-level imports in bench/gemm_ops.py instead of method-level
- Rename test method: test_flydsl_preshuffle_gemm -> test_gemm
- Move test imports to setUpClass instead of per-test
